+++ HarpGuru/HarpFace/HarpCell/components/index.test.tsx
@@ -1,3 +1,4 @@
+import { useGlobal } from 'reactn'
 import React from 'react'
 import { PitchIds, DegreeIds } from 'harpstrata'
 import { render, fireEvent } from '@testing-library/react-native'
@@ -10,9 +11,12 @@ import { DisplayModes } from '../../../types'
 
 import { HarpCell } from './index'
 
+jest.mock('reactn')
+const mockUseGlobal = useGlobal as jest.Mock
+
 test('A component is rendered with the Degree or Pitch value in its text view depending on the DisplayMode selected', () => {
+  mockUseGlobal.mockReturnValue([inactiveCellsHarpFaceProps.activeHarpStrata])
   const harpCellProps = {
-    ...inactiveCellsHarpFaceProps,
     activeDisplayMode: DisplayModes.Degree,
     yxCoord: [3, 0] as [3, 0],
   }
@@ -28,8 +32,8 @@ test('A component is rendered with the Degree or Pitch value in its text view de
 })
 
 test('A component is rendered with an a11y role of button', () => {
+  mockUseGlobal.mockReturnValue([inactiveCellsHarpFaceProps.activeHarpStrata])
   const harpCellProps = {
-    ...inactiveCellsHarpFaceProps,
     activeDisplayMode: DisplayModes.Degree,
     yxCoord: [3, 0] as [3, 0],
   }
@@ -39,8 +43,8 @@ test('A component is rendered with an a11y role of button', () => {
 })
 
 test('A component is rendered without an a11y role of button if it has no content', () => {
+  mockUseGlobal.mockReturnValue([inactiveCellsHarpFaceProps.activeHarpStrata])
   const harpCellProps = {
-    ...inactiveCellsHarpFaceProps,
     activeDisplayMode: DisplayModes.Degree,
     yxCoord: [0, 0] as [0, 0],
   }
@@ -83,11 +87,9 @@ test.skip('A press of the componenet results in toggled active ids in the harpst
 })
 
 test('A snapshot of a populated cell', () => {
-  const setActiveHarpStrata = jest.fn()
+  mockUseGlobal.mockReturnValue([inactiveCellsHarpFaceProps.activeHarpStrata])
   const harpCellProps = {
-    ...inactiveCellsHarpFaceProps,
     activeDisplayMode: DisplayModes.Degree,
-    setActiveHarpStrata,
     yxCoord: [3, 0] as [3, 0],
   }
 
@@ -96,11 +98,9 @@ test('A snapshot of a populated cell', () => {
 })
 
 test('A snapshot of an active cell', () => {
-  const setActiveHarpStrata = jest.fn()
+  mockUseGlobal.mockReturnValue([activeCellsHarpFaceProps.activeHarpStrata])
   const harpCellProps = {
-    ...activeCellsHarpFaceProps,
     activeDisplayMode: DisplayModes.Degree,
-    setActiveHarpStrata,
     yxCoord: [3, 0] as [3, 0],
   }
 
@@ -109,11 +109,9 @@ test('A snapshot of an active cell', () => {
 })
 
 test('A snapshot of an empty cell', () => {
-  const setActiveHarpStrata = jest.fn()
+  mockUseGlobal.mockReturnValue([inactiveCellsHarpFaceProps.activeHarpStrata])
   const harpFaceProps = {
-    ...inactiveCellsHarpFaceProps,
     activeDisplayMode: DisplayModes.Degree,
-    setActiveHarpStrata,
     yxCoord: [0, 0] as [0, 0],
   }