+++ HarpGuru/HarpFace/HarpCell/components/index.tsx
@@ -1,4 +1,4 @@
-import React, { useDispatch } from 'reactn'
+import { useGlobal, useDispatch } from 'reactn'
 import {
   TapGestureHandler,
   TapGestureHandlerStateChangeEvent,
@@ -6,6 +6,7 @@ import {
   LongPressGestureHandler,
 } from 'react-native-gesture-handler'
 import { Text, View } from 'react-native'
+import React from 'react'
 
 import { HarpCellProps } from '../types'
 import { toggleDegreeIdInHarpStrata } from '../toggleDegreeIdInHarpStrata'
@@ -15,17 +16,22 @@ import { getDisplayValue } from '../getDisplayValue'
 import { analysePosition } from '../analysePosition'
 
 export const HarpCell = (props: HarpCellProps): React.ReactElement => {
-  const { setActiveHarpStrata, activeHarpStrata } = props
-  const positionFacts = analysePosition(props)
+  const [activeHarpStrata, setActiveHarpStrata] = useGlobal('activeHarpStrata')
+  const harpCellProps = {
+    ...props,
+    activeHarpStrata,
+    setActiveHarpStrata,
+  }
+  const positionFacts = analysePosition(harpCellProps)
   const { thisDegree, thisPitch } = positionFacts
   const { id: degreeId } = thisDegree || { id: undefined }
   const { id: pitchId } = thisPitch || { id: undefined }
 
   const quizAnswerGiven = useDispatch('quizAnswerGiven')
 
-  const displayValue = getDisplayValue(props)
+  const displayValue = getDisplayValue(harpCellProps)
 
-  const styles = getStyles(props)
+  const styles = getStyles(harpCellProps)
 
   const handleTapStateChange = ({
     nativeEvent,