+++ HarpGuru/HarpFace/HarpCell/types/index.ts
@@ -1,15 +1,12 @@
 import type { TextStyle, ViewStyle } from 'react-native'
-import type { HarpStrata } from 'harpstrata'
 
-import type { SetActiveHarpStrata, DisplayModes } from '../../../types'
+import type { DisplayModes } from '../../../types'
 
 export type Coord = number
 
 export type YXCoord = [Coord, Coord]
 
 export type HarpCellProps = {
-  readonly activeHarpStrata: HarpStrata
-  readonly setActiveHarpStrata: SetActiveHarpStrata
   readonly activeDisplayMode: DisplayModes
   readonly yxCoord: YXCoord
 }