+++ HarpGuru/HarpFace/HarpCells/index.test.tsx
@@ -1,3 +1,4 @@
+import { useGlobal } from 'reactn'
 import { DegreeIds } from 'harpstrata'
 import { render } from '@testing-library/react-native'
 
@@ -5,6 +6,10 @@ import { harpFaceProps } from '../testResources'
 import type { HarpRowProps } from '../HarpRow'
 
 import { getHarpCells } from './index'
+
+jest.mock('reactn')
+const mockUseGlobal = useGlobal as jest.Mock
+mockUseGlobal.mockReturnValue([harpFaceProps.activeHarpStrata])
 
 test('getHarpCells returns an array of HarpCells, the length of the range supplied', () => {
   const harpRowProps: HarpRowProps = {