+++ HarpGuru/HarpFace/HarpFaceFragment/components/index.test.tsx
@@ -1,3 +1,4 @@
+import { useGlobal } from 'reactn'
 import 'react-native'
 import React from 'react'
 import { render } from '@testing-library/react-native'
@@ -5,6 +6,10 @@ import { render } from '@testing-library/react-native'
 import { harpFaceProps } from '../../testResources'
 
 import { HarpFaceFragment } from './index'
+
+jest.mock('reactn')
+const mockUseGlobal = useGlobal as jest.Mock
+mockUseGlobal.mockReturnValue([harpFaceProps.activeHarpStrata])
 
 test('A snapshot of a HarpFaceFragment', () => {
   const { container } = render(