+++ HarpGuru/HarpFace/HarpRow/types/index.ts
@@ -2,13 +2,7 @@ import type { ViewStyle } from 'react-native'
 
 import type { HarpCellProps, Coord } from '../../HarpCell'
 
-export type HarpRowProps = Pick<
-  HarpCellProps,
-  | 'activeHarpStrata'
-  | 'setActiveHarpStrata'
-  | 'activeDisplayMode'
-  | 'activeDisplayMode'
-> & {
+export type HarpRowProps = Pick<HarpCellProps, 'activeDisplayMode'> & {
   readonly yCoord: Coord
   readonly xRange: ReadonlyArray<number>
 }