+++ HarpGuru/HarpFace/components/index.tsx
@@ -1,3 +1,4 @@
+import { useGlobal } from 'reactn'
 import { View } from 'react-native'
 import React from 'react'
 
@@ -7,8 +8,12 @@ import { getHarpFaceFacts } from '../helpers'
 import { HarpFaceFragment } from '../HarpFaceFragment'
 
 export const HarpFace = (props: HarpFaceProps): React.ReactElement => {
-  const styles = getStyles(props)
-  const { octaveColumnGroups } = getHarpFaceFacts(props)
+  const [activeHarpStrata] = useGlobal('activeHarpStrata')
+  const styles = getStyles({ ...props, activeHarpStrata })
+  const { octaveColumnGroups } = getHarpFaceFacts({
+    ...props,
+    activeHarpStrata,
+  })
   const fragments = octaveColumnGroups.map((xRange, index) => {
     const harpFaceFragmentProps = {
       ...props,