+++ HarpGuru/Menus/Menus/LayoutMenu/components/index.tsx
@@ -2,17 +2,14 @@ import React from 'react'
 
 import { MenuContainer } from '../../MenuContainer'
 import { ApparatusOption, DisplayModeOption } from '../../../Options'
-import type {
-  ApparatusOptionProps,
-  DisplayModeOptionProps,
-} from '../../../Options'
+import type { DisplayModeOptionProps } from '../../../Options'
 
-export type LayoutMenuProps = ApparatusOptionProps & DisplayModeOptionProps
+export type LayoutMenuProps = DisplayModeOptionProps
 
 export const LayoutMenu = (props: LayoutMenuProps): React.ReactElement => {
   return (
     <MenuContainer>
-      <ApparatusOption {...props} />
+      <ApparatusOption />
       <DisplayModeOption {...props} />
     </MenuContainer>
   )