+++ HarpGuru/Menus/Options/ApparatusOption/components/index.test.tsx
@@ -1,10 +1,12 @@
+import { useGlobal } from 'reactn'
 import React from 'react'
 import { PitchIds, PozitionIds, ApparatusIds, getHarpStrata } from 'harpstrata'
 import { render } from '@testing-library/react-native'
 
-import { DisplayModes } from '../../../../types'
-
 import { ApparatusOption } from './index'
+
+jest.mock('reactn')
+const mockUseGlobal = useGlobal as jest.Mock
 
 const harpStrataProps = {
   apparatusId: ApparatusIds.MajorDiatonic,
@@ -16,13 +18,8 @@ const harpStrataProps = {
 const harpStrata = getHarpStrata(harpStrataProps)
 
 test('ApparatusOption renders a component with apparatus information displayed', () => {
-  const menuOptionProps = {
-    activeHarpStrata: harpStrata,
-    setActiveHarpStrata: jest.fn(),
-    activeDisplayMode: DisplayModes.Degree,
-    setActiveDisplayMode: jest.fn(),
-  }
-  const { getByText } = render(<ApparatusOption {...menuOptionProps} />)
+  mockUseGlobal.mockReturnValue([harpStrata])
+  const { getByText } = render(<ApparatusOption />)
 
   expect(getByText(ApparatusIds.MajorDiatonic)).toBeTruthy()
 })