+++ HarpGuru/Menus/Options/ApparatusOption/types/index.ts
@@ -1,8 +1,0 @@
-import type { HarpStrata } from 'harpstrata'
-
-import type { SetActiveHarpStrata } from '../../../../types'
-
-export type ApparatusOptionProps = {
-  readonly activeHarpStrata: HarpStrata
-  readonly setActiveHarpStrata: SetActiveHarpStrata
-}