+++ HarpGuru/Menus/Options/HarpKeyOption/components/index.test.tsx
@@ -1,3 +1,4 @@
+import { useGlobal } from 'reactn'
 import React from 'react'
 import { PitchIds, PozitionIds, ApparatusIds, getHarpStrata } from 'harpstrata'
 import { render } from '@testing-library/react-native'
@@ -5,6 +6,9 @@ import { render } from '@testing-library/react-native'
 import { DisplayModes } from '../../../../types'
 
 import { HarpKeyOption } from './index'
+
+jest.mock('reactn')
+const mockUseGlobal = useGlobal as jest.Mock
 
 const harpStrataProps = {
   apparatusId: ApparatusIds.MajorDiatonic,
@@ -17,11 +21,9 @@ const harpStrata = getHarpStrata(harpStrataProps)
 
 test('HarpKeyOption renders a component with harp key information displayed', () => {
   const menuOptionProps = {
-    activeHarpStrata: harpStrata,
-    setActiveHarpStrata: jest.fn(),
     activeDisplayMode: DisplayModes.Degree,
-    setActiveDisplayMode: jest.fn(),
   }
+  mockUseGlobal.mockReturnValue([harpStrata])
   const { getByText } = render(<HarpKeyOption {...menuOptionProps} />)
 
   expect(getByText(PitchIds.C)).toBeTruthy()