+++ HarpGuru/Menus/Options/RootPitchOption/types/index.ts
@@ -1,9 +1,5 @@
-import type { HarpStrata } from 'harpstrata'
-
-import type { SetActiveHarpStrata, DisplayModes } from '../../../../types'
+import type { DisplayModes } from '../../../../types'
 
 export type RootPitchOptionProps = {
-  readonly activeHarpStrata: HarpStrata
-  readonly setActiveHarpStrata: SetActiveHarpStrata
   readonly activeDisplayMode: DisplayModes
 }