+++ HarpGuru/Screens/CovariantMenuScreen/index.tsx
@@ -12,15 +12,8 @@ export const CovariantMenuScreen = (
   props: CovariantMenuScreenProps
 ): ReactElement => {
   const { onScreen } = props
-  const {
-    activeHarpStrata,
-    setActiveHarpStrata,
-    activeDisplayMode,
-    setActiveDisplayMode,
-  } = props
+  const { activeDisplayMode, setActiveDisplayMode } = props
   const covariantMenuProps = {
-    activeHarpStrata,
-    setActiveHarpStrata,
     activeDisplayMode,
     setActiveDisplayMode,
   }