+++ HarpGuru/global.d.ts
@@ -1,4 +1,5 @@
 import 'reactn'
+import type { HarpStrata } from 'harpstrata'
 
 declare module 'reactn/default' {
   export interface Reducers {
@@ -10,5 +11,6 @@ declare module 'reactn/default' {
 
   export interface State {
     counter: number
+    activeHarpStrata: HarpStrata
   }
 }