+++ HarpGuru/helpers/index.ts
@@ -1,2 +1,4 @@
 export { usePrevious } from './usePrevious'
 export { getPropsForHarpStrata } from './getPropsForHarpStrata'
+export { setGlobalReactNState } from './setGlobalReactNState'
+export type { SetActiveHarpStrata } from './setGlobalReactNState'