+++ HarpGuru/types/index.ts
@@ -1,8 +1,5 @@
-import type { HarpStrata } from 'harpstrata'
-
 export enum DisplayModes {
   Degree = 'DEGREE',
   Pitch = 'PITCH',
 }
-export type SetActiveHarpStrata = (arg0: HarpStrata) => void
 export type SetActiveDisplayMode = (arg0: DisplayModes) => void